+++ Materials/DKAC/Models/InfoModel/ChiTietInInfo.cs
@@ -25,7 +25,7 @@ namespace DKAC.Models.InfoModel
         public int? bu_in { get; set; }
         public int? bu_phat_hanh { get; set; }
 
-        public double? sl_tong { get; set; }
+        public int? sl_tong { get; set; }
 
         public int? phuong_phap_in { get; set; }
         public string phuong_phap_in_name { get; set; }
